+++ accounts/tests/test_profile.py
@@ -540,8 +540,8 @@ class ProfileTestDownloadCountFields(TestCase):
                                                      processing_state="OK", moderation_state="OK")
 
     @mock.patch('sounds.models.delete_sound_from_gaia')
-    @mock.patch('sounds.models.delete_sound_from_solr')
-    def test_download_sound_count_field_is_updated(self, delete_sound_from_solr, delete_sound_from_gaia):
+    @mock.patch('sounds.models.delete_sound_from_search_engine')
+    def test_download_sound_count_field_is_updated(self, delete_sound_from_search_engine, delete_sound_from_gaia):
         # Test downloading sounds increases the "num_sound_downloads" field
         for i in range(0, len(self.sounds)):
             Download.objects.create(user=self.user, sound=self.sounds[i], license_id=self.sounds[i].license_id)
