+++ search/management/commands/post_dirty_sounds_to_solr.py
@@ -22,7 +22,7 @@ import logging
 
 from sounds.models import Sound
 from utils.management_commands import LoggingBaseCommand
-from utils.search.search_general import add_all_sounds_to_solr, delete_sound_from_solr, check_if_sound_exists_in_solr
+from utils.search.search_general import add_all_sounds_to_search_engine, delete_sound_from_search_engine, check_if_sound_exists_in_search_egnine
 
 console_logger = logging.getLogger("console")
 
@@ -48,7 +48,7 @@ class Command(LoggingBaseCommand):
         console_logger.info("Starting posting dirty sounds to solr. %i sounds to be added/updated to the solr index"
                             % num_sounds)
 
-        num_correctly_indexed_sounds = add_all_sounds_to_solr(
+        num_correctly_indexed_sounds = add_all_sounds_to_search_engine(
             sounds_to_index, mark_index_clean=True, delete_if_existing=options['delete-if-existing'])
 
         console_logger.info("Finished posting dirty sounds to solr. %i sounds have been added/updated"
@@ -59,11 +59,11 @@ class Command(LoggingBaseCommand):
             Sound.objects.filter(is_index_dirty=True).exclude(moderation_state='OK', processing_state='OK')
         n_deleted_sounds = 0
         for sound in sounds_dirty_to_remove:
-            if check_if_sound_exists_in_solr(sound):
+            if check_if_sound_exists_in_search_egnine(sound):
                 # We need to know if the sound exists in solr so that besides deleting it (which could be accomplished
                 # by simply using delete_sound_from_solr), we know whether we have to change is_index_dirty state. If
                 # we do not change it, then we would try to delete the sound at every attempt.
-                delete_sound_from_solr(sound.id)
+                delete_sound_from_search_engine(sound.id)
                 n_deleted_sounds += 1
                 sound.is_index_dirty = False
                 sound.save()
