+++ sounds/models.py
@@ -58,7 +58,7 @@ from tickets.models import Ticket, Queue, TicketComment
 from utils.cache import invalidate_template_cache
 from utils.locations import locations_decorator
 from utils.mail import send_mail_template
-from utils.search.search_general import delete_sound_from_solr
+from utils.search.search_general import delete_sound_from_search_engine
 from utils.similarity_utilities import delete_sound_from_gaia
 from utils.sound_upload import get_csv_lines, validate_input_csv_file, bulk_describe_from_csv
 from utils.text import slugify
@@ -1119,7 +1119,7 @@ class Sound(SocialModel):
             sounds_logger.info("Send sound with id %s to queue 'analyze'" % self.id)
 
     def delete_from_indexes(self):
-        delete_sound_from_solr(self.id)
+        delete_sound_from_search_engine(self.id)
         delete_sound_from_gaia(self.id)
 
     def invalidate_template_caches(self):
