+++ sounds/tests/test_random_sound.py
@@ -50,7 +50,7 @@ class RandomSoundViewTestCase(TestCase):
 
     fixtures = ['licenses']
 
-    @mock.patch('sounds.views.get_random_sound_from_solr')
+    @mock.patch('sounds.views.get_random_sound_from_search_engine')
     def test_random_sound_view(self, random_sound):
         """ Get a sound from solr and redirect to it. """
         users, packs, sounds = create_user_and_sounds(num_sounds=1)
@@ -63,7 +63,7 @@ class RandomSoundViewTestCase(TestCase):
         self.assertEqual(response.status_code, 302)
         self.assertEqual(response.url, '/people/testuser/sounds/{}/?random_browsing=true'.format(sound.id))
 
-    @mock.patch('sounds.views.get_random_sound_from_solr')
+    @mock.patch('sounds.views.get_random_sound_from_search_engine')
     def test_random_sound_view_bad_solr(self, random_sound):
         """ Solr may send us a sound id which no longer exists (index hasn't been updated).
         In this case, use the database access """
@@ -85,7 +85,7 @@ class RandomSoundViewTestCase(TestCase):
         self.assertEqual(response.status_code, 302)
         self.assertEqual(response.url, '/people/testuser/sounds/{}/?random_browsing=true'.format(sound.id))
 
-    @mock.patch('sounds.views.get_random_sound_from_solr')
+    @mock.patch('sounds.views.get_random_sound_from_search_engine')
     def test_random_sound_view_no_solr(self, random_sound):
         """ If solr is down, get a random sound from the database and redirect to it. """
         users, packs, sounds = create_user_and_sounds(num_sounds=1)
