+++ utils/search/search_general.py
@@ -32,10 +32,9 @@ from pyparsing import ParseException
 import sounds
 from search import forms
 from search.forms import SEARCH_SORT_OPTIONS_WEB
-from utils.text import remove_control_chars
 from utils.logging_filters import get_client_ip
 from utils.search.lucene_parser import parse_query_filter_string
-from utils.search.wrapper import SearchEngine, QueryManager, SearchEngineException
+from utils.search.wrapper import SearchEngine, QueryManager, SearchEngineException, convert_to_search_engine_document
 
 search_logger = logging.getLogger("search")
 console_logger = logging.getLogger("console")
@@ -459,68 +458,15 @@ def perform_solr_query(q, current_page):
     return results.non_grouped_number_of_matches, results.facets, paginator, page, results.docs
 
 
-def convert_to_solr_document(sound):
-    document = {}
-
-    # Basic sound fields
-    keep_fields = ['username', 'created', 'is_explicit', 'avg_rating', 'is_remix', 'num_ratings', 'channels', 'md5',
-                      'was_remixed', 'original_filename', 'duration', 'type', 'id', 'num_downloads', 'filesize']
-    for key in keep_fields:
-        document[key] = getattr(sound, key)
-    document["original_filename"] = remove_control_chars(getattr(sound, "original_filename"))
-    document["description"] = remove_control_chars(getattr(sound, "description"))
-    document["tag"] = getattr(sound, "tag_array")
-    document["license"] = getattr(sound, "license_name")
-
-    if getattr(sound, "pack_id"):
-        document["pack"] = remove_control_chars(getattr(sound, "pack_name"))
-        document["grouping_pack"] = str(getattr(sound, "pack_id")) + "_" + remove_control_chars(getattr(sound, "pack_name"))
-    else:
-        document["grouping_pack"] = str(getattr(sound, "id"))
-
-    document["is_geotagged"] = False
-    if getattr(sound, "geotag_id"):
-        document["is_geotagged"] = True
-        if not math.isnan(getattr(sound, "geotag_lon")) and not math.isnan(getattr(sound, "geotag_lat")):
-            document["geotag"] = str(getattr(sound, "geotag_lon")) + " " + str(getattr(sound, "geotag_lat"))
-
-    document["bitdepth"] = getattr(sound, "bitdepth") if getattr(sound, "bitdepth") else 0
-    document["bitrate"] = getattr(sound, "bitrate") if getattr(sound, "bitrate") else 0
-    document["samplerate"] = int(getattr(sound, "samplerate")) if getattr(sound, "samplerate") else 0
-
-    document["comment"] = [remove_control_chars(comment_text) for comment_text in getattr(sound, "comments_array")]
-    document["comments"] = getattr(sound, "num_comments")
-    locations = sound.locations()
-    document["waveform_path_m"] = locations["display"]["wave"]["M"]["path"]
-    document["waveform_path_l"] = locations["display"]["wave"]["L"]["path"]
-    document["spectral_path_m"] = locations["display"]["spectral"]["M"]["path"]
-    document["spectral_path_l"] = locations["display"]["spectral"]["L"]["path"]
-    document["preview_path"] = locations["preview"]["LQ"]["mp3"]["path"]
-
-    # Audio Commons analysis
-    # NOTE: as the sound object here is the one returned by SoundManager.bulk_query_solr, it will have the Audio Commons
-    # descriptor fields under a property called 'ac_analysis'.
-    ac_analysis = getattr(sound, "ac_analysis")
-    if ac_analysis is not None:
-        # If analysis is present, index all existing analysis fields under Solr's dynamic fields "*_i", "*_d", "*_s"
-        # and "*_b" depending on the value's type. Also add Audio Commons prefix.
-        for key, value in ac_analysis.items():
-            suffix = settings.SOLR_DYNAMIC_FIELDS_SUFFIX_MAP.get(type(value), None)
-            if suffix:
-                document['{0}{1}{2}'.format(settings.AUDIOCOMMONS_DESCRIPTOR_PREFIX, key, suffix)] = value
-
-    return document
-
-
-def add_sounds_to_solr(sounds):
+def add_sounds_to_search_engine(sounds):
     search_engine = SearchEngine(settings.SOLR_URL)
-    documents = [convert_to_solr_document(s) for s in sounds]
+    documents = [convert_to_search_engine_document(s) for s in sounds]
     console_logger.info("Adding %d sounds to solr index" % len(documents))
     search_logger.info("Adding %d sounds to solr index" % len(documents))
     search_engine.add_to_index(documents)
 
 
-def add_all_sounds_to_solr(sound_queryset, slice_size=1000, mark_index_clean=False, delete_if_existing=False):
+def add_all_sounds_to_search_engine(sound_queryset, slice_size=1000, mark_index_clean=False, delete_if_existing=False):
     """
     Add all sounds from the sound_queryset to the Solr index.
     :param QuerySet sound_queryset: queryset of Sound objects.
@@ -540,8 +486,8 @@ def add_all_sounds_to_solr(sound_queryset, slice_size=1000, mark_index_clean=Fal
             sound_ids = all_sound_ids[i:i+slice_size]
             sounds_qs = sounds.models.Sound.objects.bulk_query_solr(sound_ids)
             if delete_if_existing:
-                delete_sounds_from_solr(sound_ids=sound_ids)
-            add_sounds_to_solr(sounds_qs)
+                delete_sounds_from_search_engine(sound_ids=sound_ids)
+            add_sounds_to_search_engine(sounds_qs)
 
             if mark_index_clean:
                 console_logger.info("Marking sounds as clean.")
@@ -554,7 +500,7 @@ def add_all_sounds_to_solr(sound_queryset, slice_size=1000, mark_index_clean=Fal
     return num_correctly_indexed_sounds
 
 
-def get_all_sound_ids_from_solr(limit=False):
+def get_all_sound_ids_from_search_engine(limit=False):
     search_logger.info("getting all sound ids from solr.")
     if not limit:
         limit = 99999999999999
@@ -573,14 +519,14 @@ def get_all_sound_ids_from_solr(limit=False):
     return sorted(solr_ids)
 
 
-def check_if_sound_exists_in_solr(sound):
+def check_if_sound_exists_in_search_egnine(sound):
     search_engine = SearchEngine(settings.SOLR_URL)
     response = search_engine.search(search_prepare_query(
             '', 'id:%i' % sound.id, search_prepare_sort('created asc', SEARCH_SORT_OPTIONS_WEB), 1, 1))
     return response.num_found > 0
 
 
-def get_random_sound_from_solr():
+def get_random_sound_from_search_engine():
     """ Get a random sound from solr.
     This is used for random sound browsing. We filter explicit sounds,
     but otherwise don't have any other restrictions on sound attributes
@@ -602,7 +548,7 @@ def get_random_sound_from_solr():
     return {}
 
 
-def delete_sound_from_solr(sound_id):
+def delete_sound_from_search_engine(sound_id):
     search_logger.info("deleting sound with id %d" % sound_id)
     try:
         SearchEngine(settings.SOLR_URL).remove_from_index(sound_id)
@@ -610,7 +556,7 @@ def delete_sound_from_solr(sound_id):
         search_logger.error('could not delete sound with id %s (%s).' % (sound_id, e))
 
 
-def delete_sounds_from_solr(sound_ids):
+def delete_sounds_from_search_engine(sound_ids):
     solr_max_boolean_clause = 1000  # This number is specified in solrconfig.xml
     for count, i in enumerate(range(0, len(sound_ids), solr_max_boolean_clause)):
         range_ids = sound_ids[i:i+solr_max_boolean_clause]
